dist: Skip build-tests due to test-more-bash absence
test-more-bash is at least not packaged for openSUSE. I thought of
providing an archive blob as part of the package on build.opensuse.org
but in the end decided that it's not worth it. It's very unlikely that
there is functional difference for a strict POSIX sh compliant simple
program and when we run all checks and tests within CI before accepting
any changes into the main git branch we should be safe enough.